refactor: view를 page단위로 변경, linter 적용
