+++ lib/main.dart
@@ -1,14 +1,13 @@
+import 'package:arttrip/core/app_colors.dart';
 import 'package:arttrip/core/enum.dart';
+import 'package:arttrip/features/login/login_page.dart';
+import 'package:arttrip/l10n/generated/app_localizations.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import 'core/app_colors.dart';
-import 'features/login/view/login_view.dart';
-import 'l10n/generated/app_localizations.dart';
-
 void main() {
-  WidgetsBinding widgetsBinding = WidgetsFlutterBinding.ensureInitialized();
+  var widgetsBinding = WidgetsFlutterBinding.ensureInitialized();
   FlutterNativeSplash.preserve(widgetsBinding: widgetsBinding);
   runApp(const ArtTripApp());
 }
@@ -34,7 +33,7 @@ class ArtTripApp extends StatelessWidget {
             useMaterial3: true,
             fontFamily: FontFamilyType.pretendard.fontName,
           ),
-          home: const LoginView(),
+          home: const LoginPage(),
         );
       },
     );
